Feature/transaction-CRUD (#10)
* entity-transaction

The transaction entity was established in the system and its repository was created

* Remove nulls user in model category

* Transaction-CRUD

The complete CRUD of the transaction entity was done and some validations were also put in place.

* fix-namePlaniFy

Quick fixes were made to the app name when sending emails.

* fix-bugs

Some bugs are being fixed so that the application is functional and well optimized.

---------

Co-authored-by: Jannus Alejandro Largo Zafra <jannus.largo@correounivalle.edu.co>